requirements-backend.txt
fastapi~=0.116.1
uvicorn
faster-whisper
python-multipart
pinecone~=7.3.0
python-dotenv~=1.1.1
PyMuPDF
google-genai
pydub~=0.25.1